Crate rusoto_mediapackage

AWS Elemental MediaPackage

If you're using the service, you're probably looking for MediaPackageClient and MediaPackage.

Structs

Channel
A Channel resource configuration.

ChannelCreateParameters
Configuration parameters for a new Channel.

ChannelList
A collection of Channel records.

ChannelUpdateParameters
Configuration parameters for updating an existing Channel.

CmafEncryption
A Common Media Application Format (CMAF) encryption configuration.

CmafPackage
A Common Media Application Format (CMAF) packaging configuration.

CmafPackageCreateOrUpdateParameters
A Common Media Application Format (CMAF) packaging configuration.

CreateChannelRequest
A new Channel configuration.

CreateChannelResponse

CreateOriginEndpointRequest
Configuration parameters used to create a new OriginEndpoint.

CreateOriginEndpointResponse

DashEncryption
A Dynamic Adaptive Streaming over HTTP (DASH) encryption configuration.

DashPackage
A Dynamic Adaptive Streaming over HTTP (DASH) packaging configuration.

DeleteChannelRequest

DeleteChannelResponse

DeleteOriginEndpointRequest

DeleteOriginEndpointResponse

DescribeChannelRequest

DescribeChannelResponse

DescribeOriginEndpointRequest

DescribeOriginEndpointResponse

HlsEncryption
An HTTP Live Streaming (HLS) encryption configuration.

HlsIngest
An HTTP Live Streaming (HLS) ingest resource configuration.

HlsManifest
A HTTP Live Streaming (HLS) manifest configuration.

HlsManifestCreateOrUpdateParameters
A HTTP Live Streaming (HLS) manifest configuration.

HlsPackage
An HTTP Live Streaming (HLS) packaging configuration.

IngestEndpoint
An endpoint for ingesting source content for a Channel.

ListChannelsRequest

ListChannelsResponse

ListOriginEndpointsRequest

ListOriginEndpointsResponse

MediaPackageClient
A client for the MediaPackage API.

MssEncryption
A Microsoft Smooth Streaming (MSS) encryption configuration.

MssPackage
A Microsoft Smooth Streaming (MSS) packaging configuration.

OriginEndpoint
An OriginEndpoint resource configuration.

OriginEndpointCreateParameters
Configuration parameters for a new OriginEndpoint.

OriginEndpointList
A collection of OriginEndpoint records.

OriginEndpointUpdateParameters
Configuration parameters for updating an existing OriginEndpoint.

RotateChannelCredentialsRequest

RotateChannelCredentialsResponse

SpekeKeyProvider
A configuration for accessing an external Secure Packager and Encoder Key Exchange (SPEKE) service that will provide encryption keys.

StreamSelection
A StreamSelection configuration.

UpdateChannelRequest
Configuration parameters used to update the Channel.

UpdateChannelResponse

UpdateOriginEndpointRequest
Configuration parameters used to update an existing OriginEndpoint.

UpdateOriginEndpointResponse

Enums

CreateChannelError
Errors returned by CreateChannel

CreateOriginEndpointError
Errors returned by CreateOriginEndpoint

DeleteChannelError
Errors returned by DeleteChannel

DeleteOriginEndpointError
Errors returned by DeleteOriginEndpoint

DescribeChannelError
Errors returned by DescribeChannel

DescribeOriginEndpointError
Errors returned by DescribeOriginEndpoint

ListChannelsError
Errors returned by ListChannels

ListOriginEndpointsError
Errors returned by ListOriginEndpoints

RotateChannelCredentialsError
Errors returned by RotateChannelCredentials

UpdateChannelError
Errors returned by UpdateChannel

UpdateOriginEndpointError
Errors returned by UpdateOriginEndpoint

Traits

MediaPackage
Trait representing the capabilities of the MediaPackage API. MediaPackage clients implement this trait.